chore(opentelemetry): move _span.py and _trace.py to internal package
