+++ ddtrace/internal/opentelemetry/span.py
